+++ mobile/android/android-components/components/feature/storage/src/test/java/mozilla/components/feature/storage/HistoryTrackingFeatureTest.kt
@@ -105,6 +105,11 @@ class HistoryTrackingFeatureTest {
                 return listOf()
             }
 
+            override fun getDomainSuggestion(query: String): String? {
+                fail()
+                return null
+            }
+
             override fun cleanup() {
                 fail()
             }