+++ mobile/android/android-components/docs/changelog.md
@@ -15,6 +15,10 @@ permalink: /changelog/
   * Kotlin (Stdlib: 1.2.71, Coroutines: 0.30.2)
   * GeckoView (Nightly: 65.0.20181023100123, Beta: 64.0.20181022150107, Release: 63.0.20181018182531)
 
+* **concept-storage**, **browser-storage-memory**, **browser-storage-sync**
+  * Added a `getDomainSuggestion` method to `HistoryStorage` which is intended to power awesomebar-like functionality.
+  * Added basic implementations of `getDomainSuggestion` to existing storage components.
+
 # 0.30.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.29.0...v0.30.0),